feature: Add toast alarm
