+++ src/components/FriendList/PendingFriends.js
@@ -16,6 +16,7 @@ function PendingFriends() {
       userId: user.id,
       axiosInstance,
     });
+
     setPendingFriends(pendingFriendList);
   }, []);
