+++ src/components/Town/Town.js
@@ -29,6 +29,7 @@ import Shop from "../Shop/Shop";
 import FriendProfile from "../FriendProfile/FriendProfile";
 import ShopFriendList from "../FriendList/ShopFriendList";
 import IcePalette from "./IcePalette";
+import { getMessageList } from "../../api/guestbook";
 
 const TownDiv = styled.div`
   width: 100vw;
@@ -73,6 +74,21 @@ const GiftAndItemContainer = styled.div`
     color: red;
     position: absolute;
     right: -5px;
+  }
+`;
+
+const PostBoxContainer = styled.div`
+  position: absolute;
+  top: 40%;
+  left: 10%;
+  z-index: 1;
+
+  i {
+    position: absolute;
+    top: 0px;
+    left: 57px;
+    font-size: 17px;
+    color: red;
   }
 `;
 
@@ -165,6 +181,26 @@ function Town({ socketInit }) {
     };
   }, [isLoading, id]);
 
+  useEffect(async () => {
+    const response = await getMessageList({ townId: id, axiosInstance });
+
+    const { guestBook } = response.data.result;
+    const isExistNewGuestBook = guestBook.some((msg) => !msg.isChecked);
+    setIsReceiveGuestBook(isExistNewGuestBook);
+  }, [isReceiveGuestBook]);
+
+  useEffect(() => {
+    if (isLoading) return;
+
+    socket.on(EVENTS.GET_MESSAGES, () => {
+      setIsReceiveGuestBook(true);
+    });
+
+    return () => {
+      socket.off(EVENTS.GET_MESSAGES);
+    };
+  }, [isLoading, isReceiveGuestBook]);
+
   return (
     <>
       <Header socket={socket} />
@@ -178,10 +214,12 @@ function Town({ socketInit }) {
       <TownDiv iceCount={iceCount}>
         <HostNameContainer>{hostName} 마을</HostNameContainer>
         <CokeCounter />
-        <PostBox
-          isReceiveGuestBook={isReceiveGuestBook}
-          setIsReceiveGuestBook={setIsReceiveGuestBook}
-        />
+        <PostBoxContainer>
+          <PostBox />
+          {isReceiveGuestBook && loginUser.id === id && (
+            <i className="fas fa-exclamation-circle guestBookNoti" />
+          )}
+        </PostBoxContainer>
         <IcePalette
           iceCount={iceCount}
           outItems={outItems}
